# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NOC Streaming API
  version: 0.1.0
  description: API for NOC dashboard streaming management.
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ValidationErrorResponse'
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthMeResponse:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
    LoginResponse:
      type: object
      required: [expires_in, user]
      properties:
        expires_in:
          type: integer
          minimum: 1
        user:
          $ref: '#/components/schemas/AuthMeResponse'
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ValidationIssue:
      type: object
      required: [path, message]
      properties:
        path:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      type: object
      required: [message, issues]
      properties:
        message:
          type: string
          example: validation error
        issues:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
...